Punjab_India/requirements.txt
streamlit
sqlalchemy
pandas
python-dotenv
plotly
psycopg2-binary
python-dateutil